Add files via upload
I have made mu final project on Sentiment analysis is a NLP task used to determine the sentiment behind textual data. This is a Data Science project made by me as a part of my Internship training at my College. 

Here's the description of my project
--> In context of product reviews it helps in understanding whether the feedback given by customers is positive, negative or neutral. It helps businesses gain valuable insights about customer experiences, product quality and brand perception. It helps in improving product quality and refining customer service strategies.

Code Implementation of Sentiment Analysis on Flipkart Reviews:-
1. Importing Libraries and Dataset

2. Preprocessing the Data
The next step is preprocessing the data which involves cleaning the review text and preparing the sentiment labels. We'll start by converting the reviews to lowercase and removing stopwords to make text more manageable. Then we will convert ratings (from 1 to 5) into binary sentiment labels like 1 for positive reviews (ratings 4 and 5) and 0 for negative reviews (ratings 3 and below).

 3. Visualizing the Data
Before we proceed with model making it's important to explore the dataset. We can visualize the distribution of sentiment labels and analyze the frequency of words in positive reviews.

4. Sentiment Distribution
To understand the overall sentiment distribution, we will use a bar plot to visualize the counts of positive and negative reviews.


5. Vectorizing the Text Data
Machine learning models require numerical input so we need to convert the textual reviews into numerical vectors. We will use TF-IDF (Term Frequency-Inverse Document Frequency) which helps converting these texts into vectors.

6. Model Training, Evaluation and Prediction
Now that the data is prepared we can split it into training and testing sets where 80% data is used for training and rest is used for testing. We will train a Decision Tree Classifier on the training data and evaluate its performance on the test data. We will also measure the model's accuracy and generate a confusion matrix to analyze the predictions.